style: typography variant를 style폴더로 빼고, title 컴포넌트에 variant props 추가
