+++ src/components/Common/Typography/Typography.variant.tsx
@@ -1,5 +1,4 @@
 import { css } from 'styled-components';
-import theme from '@/styles/theme';
 
 export const SIZES = {
   body1: css`
@@ -25,32 +24,5 @@ export const SIZES = {
   caption: css`
     --typography-size: 10px;
     --typography-weight: 400;
-  `
-};
-
-export const VARIANTS = {
-  default: css`
-    --typography-color: black;
-  `,
-  accent: css`
-    --typography-color: white;
-  `,
-  main: css`
-    --typography-color: ${theme.color.main};
-  `,
-  gray: css`
-    --typography-color: ${theme.color.g400};
-  `,
-  gray2: css`
-    --typography-color: ${theme.color.g500};
-  `,
-  success: css`
-    --typography-color: ${theme.color.success};
-  `,
-  warning: css`
-    --typography-color: ${theme.color.warning};
-  `,
-  error: css`
-    --typography-color: ${theme.color.error};
   `
 };
