+++ src/styles/typo.variant.tsx
@@ -1,0 +1,29 @@
+import { css } from 'styled-components';
+import theme from '@/styles/theme';
+
+export const VARIANTS = {
+  default: css`
+    --typography-color: black;
+  `,
+  accent: css`
+    --typography-color: white;
+  `,
+  main: css`
+    --typography-color: ${theme.color.main};
+  `,
+  gray: css`
+    --typography-color: ${theme.color.g400};
+  `,
+  gray2: css`
+    --typography-color: ${theme.color.g500};
+  `,
+  success: css`
+    --typography-color: ${theme.color.success};
+  `,
+  warning: css`
+    --typography-color: ${theme.color.warning};
+  `,
+  error: css`
+    --typography-color: ${theme.color.error};
+  `
+};
